// projects/ngx-apple-music/src/lib/ngx-apple-music.interface.ts
export enum MediaItemType {
    albums,
    songs,
    artists,
    musicVideos,
    playlists,
    stations
}

export enum PlaylistType {
    editorial,
    external,
    personalMix,
    replay,
    userShared
}

export enum TrackTypes {
    musicVideos,
    songs
}

export enum MediaKind {
    audio,
    video
}

export enum ContentRating {
    clean,
    explicit
}